Fix and clean up predicate serialization
